Move `Legend` and `HtmlLegend` to `charts` folder. Sort imports in `index.js`
