+++ src/charts/html-legend.js
