+++ src/charts/legend.js
