+++ src/index.js
@@ -1,23 +1,22 @@
+export * from './core/bad-argument-exception';
 export * from './core/chart-registry';
+export * from './core/config';
 export * from './core/constants';
 export * from './core/core';
-export * from './core/bad-argument-exception';
-export * from './core/invalid-state-exception';
-export * from './core/utils';
-export * from './core/logger';
-export * from './core/config';
 export * from './core/events';
 export * from './core/filters';
+export * from './core/invalid-state-exception';
+export * from './core/logger';
 export * from './core/printers';
 export * from './core/units';
-export * from './base/d3.box';
+export * from './core/utils';
 
 export * from './base/base-mixin';
 export * from './base/bubble-mixin';
+export * from './base/cap-mixin';
 export * from './base/color-mixin';
 export * from './base/coordinate-grid-mixin';
-export * from './base/cap-mixin';
-export * from './base/legend';
+export * from './base/d3.box';
 export * from './base/margin-mixin';
 export * from './base/stack-mixin';
 
@@ -32,7 +31,8 @@ export * from './charts/data-grid';
 export * from './charts/data-table';
 export * from './charts/geo-choropleth-chart';
 export * from './charts/heatmap';
-export * from './base/html-legend';
+export * from './charts/html-legend';
+export * from './charts/legend';
 export * from './charts/line-chart';
 export * from './charts/number-display';
 export * from './charts/pie-chart';
